refactor(coordbench): separate encoders from priors
