+++ docs/api/coordbench.rst
@@ -21,16 +21,6 @@ Encoders
 .. autoclass:: NeRFLocationEncoder
 .. autoclass:: SphericalHarmonicLocationEncoder
 
-Classification priors
----------------------
-
-.. autoclass:: SpatialPrior
-.. autoclass:: UniformPrior
-.. autoclass:: ClassFrequencyPrior
-.. autoclass:: GridPrior
-.. autoclass:: NearestNeighborPrior
-.. autoclass:: KDEPrior
-
 Benchmarks
 ----------
