+++ src/torchgeo_bench/conf/model/spherical-harmonics.yaml
@@ -1,4 +1,4 @@
 # Compact spherical-harmonic position encoding.
-_target_: torchgeo_bench.coordbench.baselines.SphericalHarmonicLocationEncoder
+_target_: torchgeo_bench.coordbench.models.SphericalHarmonicLocationEncoder
 name: spherical-harmonics
 degree: 3
