+++ src/torchgeo_bench/conf/model/xyz.yaml
@@ -1,3 +1,3 @@
 # XYZ position encoding; dependency-free.
-_target_: torchgeo_bench.coordbench.baselines.XYZLocationEncoder
+_target_: torchgeo_bench.coordbench.models.XYZLocationEncoder
 name: xyz
